refactor(engine): move ExecutionCache, SavedCache into reth-execution-cache
Move the concrete cache types (ExecutionCache, SavedCache, CachedStateMetrics,
PayloadExecutionCache) into the new crate. CachedStateProvider stays in
reth-engine-tree and re-exports the cache types.

Removes the CachedExecution trait — no generics needed.

Amp-Thread-ID: https://ampcode.com/threads/T-019d2125-584a-700b-84e8-b4f33b96d56f
