+++ crates/engine/execution-cache/Cargo.toml
@@ -13,8 +13,21 @@ workspace = true
 
 [dependencies]
 alloy-primitives.workspace = true
+fixed-cache = { workspace = true, features = ["stats"] }
 metrics.workspace = true
 parking_lot.workspace = true
 reth-metrics = { workspace = true, features = ["common"] }
 reth-primitives-traits = { workspace = true, features = ["std"] }
+reth-revm.workspace = true
 tracing.workspace = true
+
+[dev-dependencies]
+alloy-primitives = { workspace = true, features = ["rand"] }
+reth-revm = { workspace = true, features = ["test-utils"] }
+revm-state.workspace = true
+
+[features]
+test-utils = [
+	"reth-primitives-traits/test-utils",
+	"reth-revm/test-utils"
+]
