+++ crates/engine/tree/Cargo.toml
@@ -53,7 +53,6 @@ revm-primitives.workspace = true
 futures.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["rt", "rt-multi-thread", "sync", "macros"] }
-fixed-cache.workspace = true
 moka = { workspace = true, features = ["sync"] }
 
 # metrics
@@ -107,32 +106,33 @@ rand_08.workspace = true
 
 [features]
 test-utils = [
-    "reth-chain-state/test-utils",
-    "reth-chainspec/test-utils",
-    "reth-consensus/test-utils",
-    "reth-db/test-utils",
-    "reth-evm/test-utils",
-    "reth-network-p2p/test-utils",
-    "reth-payload-builder/test-utils",
-    "reth-primitives-traits/test-utils",
-    "reth-provider/test-utils",
-    "reth-prune-types",
-    "reth-prune-types?/test-utils",
-    "reth-revm/test-utils",
-    "reth-stages-api/test-utils",
-    "reth-stages/test-utils",
-    "reth-static-file",
-    "reth-tracing",
-    "reth-trie/test-utils",
-    "reth-trie-common/test-utils",
-    "reth-trie-db/test-utils",
-    "reth-trie-sparse/test-utils",
-    "reth-prune-types?/test-utils",
-    "reth-trie-parallel/test-utils",
-    "reth-ethereum-primitives/test-utils",
-    "reth-node-ethereum/test-utils",
-    "reth-evm-ethereum/test-utils",
-    "reth-tasks/test-utils",
+	"reth-chain-state/test-utils",
+	"reth-chainspec/test-utils",
+	"reth-consensus/test-utils",
+	"reth-db/test-utils",
+	"reth-evm/test-utils",
+	"reth-network-p2p/test-utils",
+	"reth-payload-builder/test-utils",
+	"reth-primitives-traits/test-utils",
+	"reth-provider/test-utils",
+	"reth-prune-types",
+	"reth-prune-types?/test-utils",
+	"reth-revm/test-utils",
+	"reth-stages-api/test-utils",
+	"reth-stages/test-utils",
+	"reth-static-file",
+	"reth-tracing",
+	"reth-trie/test-utils",
+	"reth-trie-common/test-utils",
+	"reth-trie-db/test-utils",
+	"reth-trie-sparse/test-utils",
+	"reth-prune-types?/test-utils",
+	"reth-trie-parallel/test-utils",
+	"reth-ethereum-primitives/test-utils",
+	"reth-node-ethereum/test-utils",
+	"reth-evm-ethereum/test-utils",
+	"reth-tasks/test-utils",
+	"reth-execution-cache/test-utils"
 ]
 trie-debug = [
     "reth-trie-sparse/trie-debug",
